# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Korner Lamp REST API documentation.",
    "title": "Korner Lamp API Documentation"
  },
  "host": "",
  "basePath": "/",
...
  "paths": {
...
    "/Category": {
      "post": {
        "tags": ["Category"],
        "summary": "Create Category",
        "produces": ["application/json"],
        "consumes": ["application/x-www-form-urlencoded"],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "description": "Name Of Category",
            "required": true,
            "type": "string"
          },
          {
            "name": "image",
            "in": "formData",
            "description": "Image",
            "required": false,
            "type": "object"
          },
          {
            "name": "tags",
            "in": "formData",
            "description": "Tags",
            "required": false,
            "type": "array",
            "items": { "type": "string" }
          },
          {
            "name": "order",
            "in": "formData",
            "description": "Order",
            "required": false,
            "type": "number"
          },
          {
            "name": "visible",
            "in": "formData",
            "description": "Visible",
            "required": false,
            "type": "boolean"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "200": {
            "description": "Category created successfully."
          },
          "500": {
            "description": "System error."
          },
          "11000": {
            "description": "Duplicate error."
          }
        }
      },
      "get": {
        "tags": ["Category"],
        "summary": "Get Categories",
        "produces": ["application/json"],
        "consumes": ["application/x-www-form-urlencoded"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "limit",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page number",
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "orderby or sortBy",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "200": {
            "description": "Category retrived successfully."
          },
          "500": {
            "description": "System error."
          }
        }
      }
    },
...
    "/Category/{_id}": {
      "get": {
        "tags": ["Category"],
        "summary": "Get A Category",
        "produces": ["application/json"],
        "consumes": ["application/x-www-form-urlencoded"],
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "description": "Id Of Category Needed",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "200": {
            "description": "Category retrived successfully."
          },
          "500": {
            "description": "System error."
          }
        }
      },
      "delete": {
        "tags": ["Category"],
        "summary": "Delete Category",
        "produces": ["application/json"],
        "consumes": ["application/x-www-form-urlencoded"],
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "description": "Id Of Cateory To Be Deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "200": {
            "description": "Category deleted successfully."
          },
          "500": {
            "description": "System error."
          }
        }
      },
      "put": {
        "tags": ["Category"],
        "summary": "Update A Category",
        "produces": ["application/json"],
        "consumes": ["application/x-www-form-urlencoded"],
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "description": "Id Of Category Needed",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Name Of Category",
            "required": true,
            "type": "string"
          },
          {
            "name": "image",
            "in": "formData",
            "description": "Image",
            "required": false,
            "type": "object"
          },
          {
            "name": "tags",
            "in": "formData",
            "description": "Tags",
            "required": false,
            "type": "array",
            "items": { "type": "string" }
          },
          {
            "name": "order",
            "in": "formData",
            "description": "Order",
            "required": false,
            "type": "number"
          },
          {
            "name": "visible",
            "in": "formData",
            "description": "Visible",
            "required": false,
            "type": "boolean"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "200": {
            "description": "Category edited successfully."
          },
          "500": {
            "description": "System error."
          }
        }
      }
    },
    "/SubCategory": {
      "post": {
        "tags": ["SubCategory"],
        "summary": "Create Sub Category",
        "produces": ["application/json"],
        "consumes": ["application/x-www-form-urlencoded"],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "description": "Name Of Sub Category",
            "required": true,
            "type": "string"
          },
          {
            "name": "image",
            "in": "formData",
            "description": "Image",
            "required": false,
            "type": "object"
          },
          {
            "name": "tags",
            "in": "formData",
            "description": "Tags",
            "required": false,
            "type": "array",
            "items": { "type": "string" }
          },
          {
            "name": "order",
            "in": "formData",
            "description": "Order",
            "required": false,
            "type": "number"
          },
          {
            "name": "visible",
            "in": "formData",
            "description": "Visible",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "parent",
            "in": "formData",
            "description": "Parent Category",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "200": {
            "description": "Sub Category created successfully."
          },
          "500": {
            "description": "System error."
          },
          "11000": {
            "description": "Duplicate error."
          }
        }
      },
      "get": {
        "tags": ["SubCategory"],
        "summary": "Get Sub Categories",
        "produces": ["application/json"],
        "consumes": ["application/x-www-form-urlencoded"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "limit",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page number",
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "orderby or sortBy",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "200": {
            "description": "Sub Category retrived successfully."
          },
          "500": {
            "description": "System error."
          }
        }
      }
    },
    "/SubCategory/{_id}": {
      "get": {
        "tags": ["SubCategory"],
        "summary": "Get A Sub Category",
        "produces": ["application/json"],
        "consumes": ["application/x-www-form-urlencoded"],
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "description": "Id Of Sub Category Needed",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "200": {
            "description": "SUb Category retrived successfully."
          },
          "500": {
            "description": "System error."
          }
        }
      },
      "delete": {
        "tags": ["SubCategory"],
        "summary": "Delete Sub Category",
        "produces": ["application/json"],
        "consumes": ["application/x-www-form-urlencoded"],
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "description": "Id Of Cateory To Be Deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "200": {
            "description": "Sub Category deleted successfully."
          },
          "500": {
            "description": "System error."
          }
        }
      },
      "put": {
        "tags": ["SubCategory"],
        "summary": "Update A Sub Category",
        "produces": ["application/json"],
        "consumes": ["application/x-www-form-urlencoded"],
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "description": "Id Of SUb Category Needed",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Name Of Sub Category",
            "required": true,
            "type": "string"
          },
          {
            "name": "image",
            "in": "formData",
            "description": "Image",
            "required": false,
            "type": "object"
          },
          {
            "name": "tags",
            "in": "formData",
            "description": "Tags",
            "required": false,
            "type": "array",
            "items": { "type": "string" }
          },
          {
            "name": "order",
            "in": "formData",
            "description": "Order",
            "required": false,
            "type": "number"
          },
          {
            "name": "visible",
            "in": "formData",
            "description": "Visible",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "parent",
            "in": "formData",
            "description": "Parent Category",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "200": {
            "description": "Sub Category edited successfully."
          },
          "500": {
            "description": "System error."
          }
        }
      }
    },
...
  },
...
}